refactor(parser): move DotDictFormatter to a separate file
